+++ internal/pb/catalog.pb.go
@@ -185,7 +185,7 @@ func (x *Shape) GetDim3() int32 {
 // Header of an image (slice of the cube)
 // It describes the image, the underlying datasets and the way to recreate it from the array of byte :
 // 1. Append ImageHeader.data and ImageChunk.data from part=0 to part=nb_parts-1
-// 2. If compression=True, decompresss the array of bytes using deflate
+// 2. If compression=True, decompress the array of bytes using deflate
 // 3. Cast the result to the dtype using byteOrder
 // 4. Reshape the result
 type ImageHeader struct {
@@ -430,7 +430,7 @@ type GetCubeRequest struct {
 	Crs              string                         `protobuf:"bytes,4,opt,name=crs,proto3" json:"crs,omitempty"`                                                                    // Coordinates Reference System of the output images (images will be reprojected on the fly if necessary)
 	PixToCrs         *GeoTransform                  `protobuf:"bytes,5,opt,name=pix_to_crs,json=pixToCrs,proto3" json:"pix_to_crs,omitempty"`                                        // GeoTransform of the requested cube (images will be rescaled on the fly if necessary)
 	Size             *Size                          `protobuf:"bytes,6,opt,name=size,proto3" json:"size,omitempty"`                                                                  // Shape of the output images
-	CompressionLevel int32                          `protobuf:"varint,7,opt,name=compression_level,json=compressionLevel,proto3" json:"compression_level,omitempty"`                 // Define a level of compression to speed up the transfer, values: -2 to 9 (-2: Huffman only, -1:default, 0: no compression, 1->9: level of compression from the fastest to the best compression). The data is compressed by the server and decompressed by the Client. Use -2 or 0 if the bandwidth is not limited.
+	CompressionLevel int32                          `protobuf:"varint,7,opt,name=compression_level,json=compressionLevel,proto3" json:"compression_level,omitempty"`                 // Define a level of compression to speed up the transfer, values: -3 to 9 (-2: Huffman only, -1:default, 0->9: level of compression from the fastest to the best compression, -3: disable the compression). The data is compressed by the server and decompressed by the Client. Use -3 or -2 if the bandwidth is not limited. 0 is level 0 of DEFLATE (thus, it must be decompressed by DEFLATE even though the data is not compressed). If the client can support -3, 0 is useless.
 	HeadersOnly      bool                           `protobuf:"varint,8,opt,name=headers_only,json=headersOnly,proto3" json:"headers_only,omitempty"`                                // Only returns headers (including all metadatas on datasets)
 	Format           FileFormat                     `protobuf:"varint,9,opt,name=format,proto3,enum=geocube.FileFormat" json:"format,omitempty"`                                     // Format of the output images
 	ResamplingAlg    Resampling                     `protobuf:"varint,10,opt,name=resampling_alg,json=resamplingAlg,proto3,enum=geocube.Resampling" json:"resampling_alg,omitempty"` // Resampling algorithm used for reprojecion. If undefined, the default resampling algorithm associated to the variable is used.